+++ src/equip/cache/loadout1.txt
@@ -1,0 +1,1 @@
+{Space} {Space} d {Space} d d {Space} d d d {Space} {Space} s {Space} s d {Space} d {Space} {Space}
